Move SaleSummaryBar and InfoItem to the components folder
